// packages/client/platform/http/src-web/bindings.ts
// This file has been generated by TauRPC. Do not edit this file manually.


import { createTauRPCProxy as createProxy, type InferCommandOutput, type TauRpcResult } from 'taurpc'
export type BasicAuth = {
	username: string,
	password: string,
};

export type BodyChunk = {
	bytes: number[],
	done: boolean,
};

export type ClearCookiesConfig = {
	instanceKey: string,
};

export type ClientConfig = {
	connectTimeout: number | null,
	maxRedirections: number | null,
	proxy: Proxy | null,
	danger: DangerousSettings | null,
	userAgent: string | null,
	instanceKey: string,
};

export type ContentConfig = {
	method: string,
	url: string,
	headers: ([string, string])[],
	data: number[] | null,
	client: ClientConfig,
};

export type CookieEntry = {
	domain: string,
	name: string,
	value: string,
};

export type DangerousSettings = {
	acceptInvalidCerts: boolean,
	acceptInvalidHostnames: boolean,
};

export type DeleteCookieConfig = {
	url: string,
	path: string | null,
	name: string,
	instanceKey: string,
};

export type Error = string;

export type FetchResponse = {
	status: number,
	statusText: string,
	headers: ([string, string])[],
	url: string,
	rid: number,
};

export type GetAllCookiesConfig = {
	instanceKey: string,
};

export type GetAllDomainCookiesConfig = {
	url: string,
	instanceKey: string,
};

export type GetCookieConfig = {
	url: string,
	name: string,
	instanceKey: string,
};

export type Proxy = {
	all: UrlOrConfig | null,
	http: UrlOrConfig | null,
	https: UrlOrConfig | null,
};

export type ProxyConfig = {
	url: string,
	basicAuth: BasicAuth | null,
	noProxy: string | null,
};

export type Result<T, E> = {
	ok: T,
	err: E,
};

export type SetCookieConfig = {
	url: string,
	content: string,
	instanceKey: string,
};

export type UrlOrConfig = string | ProxyConfig;
const ARGS_MAP = {"http":{"clear_cookie":["config"],"delete_cookie":["config"],"fetch":["content_config"],"fetch_cancel":["rid"],"fetch_cancel_body":["rid"],"fetch_read_body":["rid"],"fetch_send":["rid"],"get_all_cookies":["config"],"get_all_domain_cookies":["config"],"get_cookie":["config"],"prepare_requester":["client"],"set_cookie":["config"]}};

const RESULT_MAP = {"http":{"clear_cookie":true,"delete_cookie":true,"fetch":true,"fetch_cancel":true,"fetch_cancel_body":true,"fetch_read_body":true,"fetch_send":true,"get_all_cookies":true,"get_all_domain_cookies":true,"get_cookie":true,"prepare_requester":false,"set_cookie":true}};

export type Router = {
	http: {
		clear_cookie: (config: ClearCookiesConfig) => Promise<null>,
		delete_cookie: (config: DeleteCookieConfig) => Promise<boolean>,
		fetch: (contentConfig: ContentConfig) => Promise<number>,
		fetch_cancel: (rid: number) => Promise<null>,
		fetch_cancel_body: (rid: number) => Promise<null>,
		fetch_read_body: (rid: number) => Promise<BodyChunk>,
		fetch_send: (rid: number) => Promise<FetchResponse>,
		get_all_cookies: (config: GetAllCookiesConfig) => Promise<CookieEntry[]>,
		get_all_domain_cookies: (config: GetAllDomainCookiesConfig) => Promise<CookieEntry[]>,
		get_cookie: (config: GetCookieConfig) => Promise<string | null>,
		prepare_requester: (client: ClientConfig) => Promise<void>,
		set_cookie: (config: SetCookieConfig) => Promise<null>,
	},
};
export const createTauRPCProxy = () => createProxy<Router>({
  argsMap: ARGS_MAP,
  resultMap: RESULT_MAP,
  errorHandling: "throw",
})

// export const createTauRPCProxy = () => createProxy<Router>(ARGS_MAP)
export type { InferCommandOutput, TauRpcResult }